requirements.txt
numpy
tensorflow
fastapi
uvicorn
pillow
uvicorn
python-multipart
